Fixes enter_raw_repl failures
This fixes 2 problems that can occur. The first issue is when you used ampy the first time after plugging in a device. The fix for this requires that the reset button be pressed after plugging in the device before ampy runs.

The second problem is the number of CTRL+C combinations needed to get a program to terminate properly. With threads being available it appears as tho a CTRL+C is needed for each thread that is running.